backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
